refactor(aead): move Decrypt into its own module to mirror Encrypt
The decrypt side had `Decrypt` — the trait a plaintext type implements to
describe how it decodes itself — living alongside `Decipher`, the trait a
cipher backend implements to drive that decoding. The encrypt side already
keeps the two apart (`encrypt` vs `cipher`), so the asymmetry made the pairing
harder to see than it is.

`Decrypt` and its impls now sit in `decrypt`, opposite `encrypt`, leaving
`decipher` as the mirror of `cipher`. Public paths are unchanged: everything
is re-exported flat from the crate root as before.
